+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # [NEXT]
 - Adds param `backgroundColor` in `BonfireWidget`.
 - create `BarLifeComponent`.
+- Adds `UseBarLife` mixin.
 
 # [2.10.10]
 - Update Flame to `1.4.0`.